+++ bundle-with-policy.js
@@ -1,0 +1,3 @@
+export * from './bundle.js';
+import './default-policy.js';
+export const policy = 'trustedTypes' in globalThis ? trustedTypes.defaultPolicy : null;